chore: simplify ImplicitPackagesResolver to simply provide versions
