+++ src/Uno.Sdk/Tasks/ImplicitPackagesResolver.cs
@@ -1,296 +1,411 @@
+using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Linq;
+using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Text;
+using System.Text.RegularExpressions;
+using Microsoft.Build.Framework;
+using Microsoft.Build.Utilities;
+using Uno.Sdk.Models;
+using Uno.Sdk.Services;
+
 namespace Uno.Sdk.Tasks;
 
-public sealed class ImplicitPackagesResolver_v0 : ImplicitPackagesResolverBase
+public sealed class ImplicitPackagesResolver_v0 : Task
 {
-	protected override void ExecuteInternal()
+	private readonly List<string> _existingReferences = [];
+	private PackageManifest? _manifest;
+	private NuGetVersion? _unoVersion;
+
+	public bool SdkDebugging { get; set; }
+
+	public bool SingleProject { get; set; }
+
+	[Required]
+	public string OutputType { get; set; } = null!;
+
+	public bool IsPackable { get; set; }
+
+	public bool Optimize { get; set; }
+
+	[Required]
+	public string IntermediateOutput { get; set; } = null!;
+
+	public string? UnoFeatures { get; set; }
+
+	[Required]
+	public string TargetFramework { get; set; } = null!;
+
+	private string TargetFrameworkVersion = null!;
+
+	private string TargetRuntime = null!;
+
+	[Required]
+	public string ProjectName { get; set; } = null!;
+
+	public string? UnoExtensionsVersion { get; set; }
+
+	public string? UnoToolkitVersion { get; set; }
+
+	public string? UnoThemesVersion { get; set; }
+
+	public string? UnoCSharpMarkupVersion { get; set; }
+
+	public string? MauiVersion { get; set; }
+
+	public string? SkiaSharpVersion { get; set; }
+
+	public string? SvgSkiaVersion { get; set; }
+
+	public string? UnoLoggingVersion { get; set; }
+
+	public string? WindowsCompatibilityVersion { get; set; }
+
+	public string? UnoWasmBootstrapVersion { get; set; }
+
+	public string? UnoUniversalImageLoaderVersion { get; set; }
+
+	public string? AndroidMaterialVersion { get; set; }
+
+	public string? AndroidXLegacySupportV4Version { get; set; }
+
+	public string? AndroidXAppCompatVersion { get; set; }
+
+	public string? AndroidXRecyclerViewVersion { get; set; }
+
+	public string? AndroidXActivityVersion { get; set; }
+
+	public string? AndroidXBrowserVersion { get; set; }
+
+	public string? AndroidXSwipeRefreshLayoutVersion { get; set; }
+
+	public string? UnoResizetizerVersion { get; set; }
+
+	public string? MicrosoftLoggingVersion { get; set; }
+
+	public string? WinAppSdkVersion { get; set; }
+
+	public string? WinAppSdkBuildToolsVersion { get; set; }
+
+	public string? UnoCoreLoggingSingletonVersion { get; set; }
+
+	public string? UnoDspTasksVersion { get; set; }
+
+	public string? CommunityToolkitMvvmVersion { get; set; }
+
+	public string? PrismVersion { get; set; }
+
+	public string? AndroidXNavigationVersion { get; set; }
+
+	public string? AndroidXCollectionVersion { get; set; }
+
+	public string? MicrosoftIdentityClientVersion { get; set; }
+
+	public ITaskItem[] ImplicitPackageReferences { get; set; } = [];
+
+	public ITaskItem[] PackageReferences { get; set; } = [];
+
+	public ITaskItem[] PackageVersions { get; set; } = [];
+
+	private readonly List<PackageReference> _implicitPackages = [];
+	[Output]
+	public ITaskItem[] ImplicitPackages => [.. _implicitPackages.Distinct()
+		.Select(x => x.ToTaskItem())];
+
+	[Output]
+	public ITaskItem[] RemovePackageVersions =>
+		PackageVersions.Where(x =>
+			_implicitPackages.Any(ip => ip.PackageId == x.ItemSpec)).ToArray();
+
+	private UnoFeature[] _unoFeatures = [];
+	public sealed override bool Execute()
 	{
-		AddUnoCorePackages();
-		AddUnoCSharpMarkup();
-		AddUnoExtensionsPackages();
-		AddUnoToolkitPackages();
-		AddUnoThemes();
-		AddPrism();
-		AddPackageForFeature(UnoFeature.Dsp, "Uno.Dsp.Tasks", UnoDspTasksVersion);
-		AddPackageForFeature(UnoFeature.Mvvm, "CommunityToolkit.Mvvm", CommunityToolkitMvvmVersion);
-	}
-
-	private void AddUnoCorePackages()
-	{
-		AddPackage("Uno.WinUI", null);
-
-		if (!IsPackable)
+		try
 		{
-			AddPackage("Uno.UI.Adapter.Microsoft.Extensions.Logging", null);
-			AddPackage("Uno.Resizetizer", UnoResizetizerVersion);
-			AddPackage("Microsoft.Extensions.Logging.Console", MicrosoftLoggingVersion);
-		}
-
-		AddPackageForFeature(UnoFeature.Maps, "Uno.WinUI.Maps", null);
-		AddPackageForFeature(UnoFeature.Foldable, "Uno.WinUI.Foldable", null);
-
-		var devServerExcludes = Optimize ? "all" : null;
-		AddPackageWhen(IsExecutable, "Uno.WinUI.DevServer", null, devServerExcludes);
-
-		if (TargetRuntime != UnoTarget.Windows)
-		{
-			if (HasFeature(UnoFeature.Material) || HasFeature(UnoFeature.Lottie) || HasFeature(UnoFeature.Cupertino))
+			if (TargetFramework.Contains('-'))
 			{
-				AddPackage("Uno.WinUI.Lottie", null);
-				AddPackageWhen(TargetRuntime != UnoTarget.Wasm && !IsLegacyWasmHead(), "SkiaSharp.Skottie", SkiaSharpVersion);
-			}
-
-			if (HasFeature(UnoFeature.Material) || HasFeature(UnoFeature.Skia) || HasFeature(UnoFeature.Lottie) || HasFeature(UnoFeature.Svg) || HasFeature(UnoFeature.Cupertino))
-			{
-				AddPackage("SkiaSharp.Views.Uno.WinUI", SkiaSharpVersion);
-			}
-
-			AddPackageForFeatureWhen(IsExecutable, UnoFeature.Svg, "Uno.WinUI.Svg", null);
-
-			if (TargetRuntime == UnoTarget.Wasm || IsLegacyWasmHead())
-			{
-				AddPackage("Uno.WinUI.WebAssembly", null);
+				var frameworkParts = TargetFramework.Split('-');
+				TargetFrameworkVersion = frameworkParts[0];
+				var runtime = frameworkParts[1].ToLowerInvariant();
+				if (runtime.Contains("windows"))
+				{
+					TargetRuntime = runtime.StartsWith(UnoTarget.Windows, StringComparison.InvariantCultureIgnoreCase) ? UnoTarget.Windows : UnoTarget.SkiaWpf;
+				}
+				else
+				{
+					TargetRuntime = runtime;
+				}
 			}
 			else
 			{
-				AddPackageForFeatureWhen(IsExecutable, UnoFeature.Svg, "Svg.Skia", SvgSkiaVersion);
+				TargetFrameworkVersion = TargetFramework;
+				TargetRuntime = UnoTarget.Reference;
+				if (ProjectName.EndsWith("Skia.Gtk", StringComparison.InvariantCultureIgnoreCase))
+				{
+					TargetRuntime = UnoTarget.SkiaGtk;
+				}
+				else if (ProjectName.EndsWith("Skia.WPF", StringComparison.InvariantCultureIgnoreCase))
+				{
+					TargetRuntime = UnoTarget.SkiaWpf;
+				}
+				else if (ProjectName.EndsWith("Skia.Linux.FrameBuffer", StringComparison.InvariantCultureIgnoreCase))
+				{
+					TargetRuntime = UnoTarget.SkiaLinuxFramebuffer;
+				}
 			}
 
-			AddCorePackagesForExecutable();
-		}
-		else
-		{
-			AddPackage("Microsoft.WindowsAppSDK", WinAppSdkVersion);
-			AddPackage("Microsoft.Windows.SDK.BuildTools", WinAppSdkBuildToolsVersion);
-			AddPackageWhen(IsExecutable, "Uno.Core.Extensions.Logging.Singleton", UnoCoreLoggingSingletonVersion);
-
-			// Match the conditions to Uno so that we have the reference across all targets
-			if (HasFeature(UnoFeature.Material) || HasFeature(UnoFeature.Skia) || HasFeature(UnoFeature.Lottie) || HasFeature(UnoFeature.Svg) || HasFeature(UnoFeature.Cupertino))
+			_manifest = new PackageManifest(Log, TargetFrameworkVersion);
+			if (NuGetVersion.TryParse(_manifest.UnoVersion, out var unoVersion))
 			{
-				// NOTE: This will change to come through transitively in SkiaSharp 3
-				AddPackage("SkiaSharp.Views.WinUI", SkiaSharpVersion);
+				_unoVersion = unoVersion;
 			}
-		}
-	}
-
-	private void AddCorePackagesForExecutable()
-	{
-		if (!IsExecutable)
-		{
-			Debug("Skipping Core packages for Library build.");
-			return;
-		}
-
-		if (TargetRuntime == UnoTarget.Android)
-		{
-			if (!HasFeature(UnoFeature.MauiEmbedding))
+			else
 			{
-				AddPackage("Xamarin.Google.Android.Material", AndroidMaterialVersion);
+				throw new InvalidOperationException("Unable to parse UnoVersion from the Package Manifest.");
 			}
 
-			AddPackage("Uno.UniversalImageLoader", UnoUniversalImageLoaderVersion);
-			AddPackage("Xamarin.AndroidX.Legacy.Support.V4", AndroidXLegacySupportV4Version);
-			AddPackage("Xamarin.AndroidX.AppCompat", AndroidXAppCompatVersion);
-			AddPackage("Xamarin.AndroidX.RecyclerView", AndroidXRecyclerViewVersion);
-			AddPackage("Xamarin.AndroidX.Activity", AndroidXActivityVersion);
-			AddPackage("Xamarin.AndroidX.Browser", AndroidXBrowserVersion);
-			AddPackage("Xamarin.AndroidX.SwipeRefreshLayout", AndroidXSwipeRefreshLayoutVersion);
-		}
-		else if (TargetRuntime == UnoTarget.iOS)
-		{
-			AddPackage("Uno.Extensions.Logging.OSLog", UnoLoggingVersion);
-		}
-		else if (TargetRuntime == UnoTarget.MacCatalyst)
-		{
-			AddPackage("Uno.Extensions.Logging.OSLog", UnoLoggingVersion);
-		}
-		else if (TargetRuntime == UnoTarget.SkiaDesktop)
-		{
-			AddPackage("Uno.WinUI.Skia.Linux.FrameBuffer", null);
-			AddPackage("Uno.WinUI.Skia.MacOS", null);
-			AddPackage("Uno.WinUI.Skia.Wpf", null);
-			AddPackage("Uno.WinUI.Skia.X11", null);
-		}
-		else if (TargetRuntime == UnoTarget.SkiaGtk)
-		{
-			AddPackage("Uno.WinUI.Skia.Gtk", null);
-			AddPackage("SkiaSharp.NativeAssets.Linux", SkiaSharpVersion);
-			AddPackage("SkiaSharp.NativeAssets.macOS", SkiaSharpVersion);
-		}
-		else if (TargetRuntime == UnoTarget.SkiaLinuxFramebuffer)
-		{
-			AddPackage("Uno.WinUI.Skia.Linux.FrameBuffer", null);
-			AddPackage("SkiaSharp.NativeAssets.Linux", SkiaSharpVersion);
-		}
-		else if (TargetRuntime == UnoTarget.SkiaWpf)
-		{
-			AddPackage("Uno.WinUI.Skia.Wpf", null);
-			AddPackage("SkiaSharp.NativeAssets.Win32", SkiaSharpVersion);
-		}
-		else if (TargetRuntime == UnoTarget.Wasm || IsLegacyWasmHead())
-		{
-			AddPackageForFeature(UnoFeature.MediaElement, "Uno.WinUI.MediaPlayer.WebAssembly", null);
-			AddPackage("Microsoft.Windows.Compatibility", WindowsCompatibilityVersion);
+			_unoFeatures = GetFeatures();
+			if (Log.HasLoggedErrors)
+			{
+				return false;
+			}
 
-			AddPackage("Uno.Extensions.Logging.WebAssembly.Console", UnoLoggingVersion);
-			AddPackage("Uno.Wasm.Bootstrap", UnoWasmBootstrapVersion);
-			AddPackage("Uno.Wasm.Bootstrap.DevServer", UnoWasmBootstrapVersion);
+			// TODO: Add update from Manifest that can ship via nuget.org
+			SetupRuntimePackageManifestUpdates(_manifest);
+			foreach (var reference in ImplicitPackageReferences)
+			{
+				var packageId = reference.ItemSpec;
+				var excludeAssets = packageId == "Uno.WinUI.DevServer" && Optimize ? "all" : null;
+				AddPackage(packageId, excludeAssets);
+			}
 		}
+		catch (Exception ex)
+		{
+			Log.LogErrorFromException(ex);
+
+			if (SdkDebugging)
+			{
+				Log.LogMessage(MessageImportance.High, ex.ToString());
+			}
+		}
+
+		if (_existingReferences.Count > 0)
+		{
+			var builder = new StringBuilder();
+			builder.AppendLine("Uno Platform Implicit Package references are enabled, you should remove these references from your csproj:");
+			_existingReferences.Select(x => $"\t<PackageReference Include=\"{x}\" />")
+				.ToList()
+				.ForEach(x => builder.AppendLine(x));
+			builder.AppendLine("See https://aka.platform.uno/UNOB0009 for more information.");
+			Log.LogMessage(subcategory: null,
+				code: "UNOB0009",
+				helpKeyword: null,
+				file: null,
+				lineNumber: 0,
+				columnNumber: 0,
+				endLineNumber: 0,
+				endColumnNumber: 0,
+				MessageImportance.Normal,
+				message: builder.ToString());
+		}
+
+#if DEBUG
+		var missingImplicitPackages = PackageReferences.Where(x => !string.IsNullOrEmpty(x.GetMetadata("ProjectSystem"))
+			&& !_implicitPackages.Any(p => p.PackageId == x.ItemSpec))
+			.ToArray();
+
+		if (missingImplicitPackages.Length > 0)
+		{
+			System.Diagnostics.Debugger.Launch();
+		}
+#endif
+
+		return !Log.HasLoggedErrors;
 	}
 
-	private void AddUnoCSharpMarkup()
+	private void SetupRuntimePackageManifestUpdates(PackageManifest manifest)
 	{
-		if (!HasFeature(UnoFeature.CSharpMarkup))
-		{
-			return;
-		}
-
-		AddPackage("Uno.WinUI.Markup", UnoCSharpMarkupVersion);
-		AddPackage("Uno.Extensions.Markup.Generators", UnoCSharpMarkupVersion);
-	}
-
-	public void AddUnoExtensionsPackages()
-	{
-		var useExtensions = HasFeature(UnoFeature.Extensions);
-		AddPackageForFeature(UnoFeature.Authentication, "Uno.Extensions.Authentication.WinUI", UnoExtensionsVersion);
-
-		if (HasFeature(UnoFeature.AuthenticationMsal))
-		{
-			AddPackage("Uno.Extensions.Authentication.MSAL.WinUI", UnoExtensionsVersion);
-			AddPackage("Microsoft.Identity.Client", MicrosoftIdentityClientVersion);
-			AddPackage("Uno.WinUI.MSAL", null);
-		}
-
-		AddPackageForFeature(UnoFeature.AuthenticationOidc, "Uno.Extensions.Authentication.Oidc.WinUI", UnoExtensionsVersion);
-
-		if (useExtensions || HasFeature(UnoFeature.Configuration))
-		{
-			AddPackage("Uno.Extensions.Configuration", UnoExtensionsVersion);
-		}
-
-		if (useExtensions || HasFeature(UnoFeature.ExtensionsCore) || HasFeature(UnoFeature.ThemeService))
-		{
-			AddPackage("Uno.Extensions.Core.WinUI", UnoExtensionsVersion);
-		}
-
-		if (useExtensions || HasFeature(UnoFeature.Hosting))
-		{
-			AddPackage("Uno.Extensions.Hosting.WinUI", UnoExtensionsVersion);
-		}
-
-		if (HasFeature(UnoFeature.Http))
-		{
-			AddPackage("Uno.Extensions.Http.WinUI", UnoExtensionsVersion);
-			AddPackage("Uno.Extensions.Http.Refit", UnoExtensionsVersion);
-		}
-
-		if (HasFeature(UnoFeature.Localization))
-		{
-			AddPackage("Uno.Extensions.Localization.WinUI", UnoExtensionsVersion);
-		}
-
-		if (useExtensions || HasFeature(UnoFeature.Logging))
-		{
-			AddPackage("Uno.Extensions.Logging.WinUI", UnoExtensionsVersion);
-		}
+		// Checks any MSBuild parameters passed to the task to override the default versions from the bundled packages.json
+		manifest.UpdateManifest(PackageManifest.Group.WasmBootstrap, UnoWasmBootstrapVersion)
+			.UpdateManifest(PackageManifest.Group.OSLogging, UnoLoggingVersion)
+			.UpdateManifest(PackageManifest.Group.CoreLogging, UnoCoreLoggingSingletonVersion)
+			.UpdateManifest(PackageManifest.Group.UniversalImageLoading, UnoUniversalImageLoaderVersion)
+			.UpdateManifest(PackageManifest.Group.Dsp, UnoDspTasksVersion)
+			.UpdateManifest(PackageManifest.Group.Resizetizer, UnoResizetizerVersion)
+			.UpdateManifest(PackageManifest.Group.SkiaSharp, SkiaSharpVersion)
+			.UpdateManifest(PackageManifest.Group.SvgSkia, SvgSkiaVersion)
+			.UpdateManifest(PackageManifest.Group.WinAppSdk, WinAppSdkVersion)
+			.UpdateManifest(PackageManifest.Group.WinAppSdkBuildTools, WinAppSdkBuildToolsVersion)
+			.UpdateManifest(PackageManifest.Group.MicrosoftLoggingConsole, MicrosoftLoggingVersion)
+			.UpdateManifest(PackageManifest.Group.WindowsCompatibility, WindowsCompatibilityVersion)
+			.UpdateManifest(PackageManifest.Group.MsalClient, MicrosoftIdentityClientVersion)
+			.UpdateManifest(PackageManifest.Group.Mvvm, CommunityToolkitMvvmVersion)
+			.UpdateManifest(PackageManifest.Group.Prism, PrismVersion)
+			.UpdateManifest(PackageManifest.Group.AndroidMaterial, AndroidMaterialVersion)
+			.UpdateManifest(PackageManifest.Group.AndroidXLegacySupportV4, AndroidXLegacySupportV4Version)
+			.UpdateManifest(PackageManifest.Group.AndroidXAppCompat, AndroidXAppCompatVersion)
+			.UpdateManifest(PackageManifest.Group.AndroidXRecyclerView, AndroidXRecyclerViewVersion)
+			.UpdateManifest(PackageManifest.Group.AndroidXActivity, AndroidXActivityVersion)
+			.UpdateManifest(PackageManifest.Group.AndroidXBrowser, AndroidXBrowserVersion)
+			.UpdateManifest(PackageManifest.Group.AndroidXSwipeRefreshLayout, AndroidXSwipeRefreshLayoutVersion)
+			.UpdateManifest(PackageManifest.Group.AndroidXNavigation, AndroidXNavigationVersion)
+			.UpdateManifest(PackageManifest.Group.AndroidXCollection, AndroidXCollectionVersion)
+			.UpdateManifest(PackageManifest.Group.CSharpMarkup, UnoCSharpMarkupVersion)
+			.UpdateManifest(PackageManifest.Group.Extensions, UnoExtensionsVersion)
+			.UpdateManifest(PackageManifest.Group.Toolkit, UnoToolkitVersion)
+			.UpdateManifest(PackageManifest.Group.Themes, UnoThemesVersion);
 
 		if (HasFeature(UnoFeature.MauiEmbedding))
 		{
-			AddPackage("Uno.Extensions.Maui.WinUI", UnoExtensionsVersion);
-			AddPackageForFeature(UnoFeature.CSharpMarkup, "Uno.Extensions.Maui.WinUI.Markup", UnoExtensionsVersion);
-
-			AddPackage("Microsoft.Maui.Controls", MauiVersion);
-			AddPackage("Microsoft.Maui.Controls.Compatibility", MauiVersion);
-			AddPackage("Microsoft.Maui.Graphics", MauiVersion);
-
-			if (SingleProject)
-			{
-				AddPackage("Microsoft.Maui.Controls.Build.Tasks", MauiVersion, "all");
-			}
-
-			if (TargetRuntime == UnoTarget.Android)
-			{
-				AddPackage("Xamarin.AndroidX.Navigation.UI", AndroidXNavigationVersion);
-				AddPackage("Xamarin.AndroidX.Navigation.Fragment", AndroidXNavigationVersion);
-				AddPackage("Xamarin.AndroidX.Navigation.Runtime", AndroidXNavigationVersion);
-				AddPackage("Xamarin.AndroidX.Navigation.Common", AndroidXNavigationVersion);
-				AddPackage("Xamarin.AndroidX.Collection", AndroidXCollectionVersion);
-				AddPackage("Xamarin.AndroidX.Collection.Ktx", AndroidXCollectionVersion);
-			}
-		}
-
-		if (HasFeature(UnoFeature.Navigation) && !HasFeature(UnoFeature.Prism))
-		{
-			AddPackage("Uno.Extensions.Navigation.WinUI", UnoExtensionsVersion);
-			AddPackageForFeature(UnoFeature.CSharpMarkup, "Uno.Extensions.Navigation.WinUI.Markup", UnoExtensionsVersion);
-			AddPackageForFeature(UnoFeature.Toolkit, "Uno.Extensions.Navigation.Toolkit.WinUI", UnoExtensionsVersion);
-		}
-
-		if (HasFeature(UnoFeature.MVUX))
-		{
-			AddPackage("Uno.Extensions.Reactive.WinUI", UnoExtensionsVersion);
-			AddPackage("Uno.Extensions.Reactive.Messaging", UnoExtensionsVersion);
-			AddPackageForFeature(UnoFeature.CSharpMarkup, "Uno.Extensions.Reactive.WinUI.Markup", UnoExtensionsVersion);
-		}
-
-		if (HasFeature(UnoFeature.Serialization))
-		{
-			AddPackage("Uno.Extensions.Serialization.Http", UnoExtensionsVersion);
-			AddPackage("Uno.Extensions.Serialization.Refit", UnoExtensionsVersion);
-		}
-
-		if (HasFeature(UnoFeature.LoggingSerilog))
-		{
-			AddPackage("Uno.Extensions.Logging.Serilog", UnoExtensionsVersion);
-		}
-
-		if (HasFeature(UnoFeature.Storage))
-		{
-			AddPackage("Uno.Extensions.Storage.WinUI", UnoExtensionsVersion);
+			manifest.UpdateManifest(PackageManifest.Group.Maui, MauiVersion);
 		}
 	}
 
-	public void AddUnoToolkitPackages()
+	private bool HasFeature(UnoFeature feature) =>
+		_unoFeatures.Any(x => x == feature);
+
+	private UnoFeature[] GetFeatures()
 	{
-		if (!HasFeature(UnoFeature.Toolkit))
+		if (string.IsNullOrEmpty(UnoFeatures))
 		{
+			Debug("UnoFeatures was provided as an empty or null value.");
+			return [];
+		}
+
+		var features = Regex.Replace(UnoFeatures, @"\s", string.Empty)
+			.Replace(',', ';');
+		if (string.IsNullOrEmpty(features))
+		{
+			Debug("No UnoFeatures were provided.");
+			return [];
+		}
+
+		var unoFeatures = features.Split(';')
+			.Select(x => x.Trim()) // sanity check
+			.Where(x => !string.IsNullOrEmpty(x))
+			.Distinct()
+			.Select(ParseFeature)
+			.Where(x => x != UnoFeature.Invalid)
+			.ToArray();
+
+		Debug("Found {0} UnoFeatures for platform {1}.", unoFeatures.Length, TargetFramework ?? "Default");
+		return unoFeatures;
+	}
+
+	private UnoFeature ParseFeature(string feature)
+	{
+		if (Enum.TryParse<UnoFeature>(feature, true, out var unoFeature))
+		{
+			Debug("Parsed UnoFeature: '{0}'.", feature);
+			ValidateFeature(unoFeature);
+			return unoFeature;
+		}
+
+		Log.LogWarning($"Unable to parse '{feature}' to a known Uno Feature.");
+		return UnoFeature.Invalid;
+	}
+
+	public void ValidateFeature(UnoFeature feature)
+	{
+		var area = typeof(UnoFeature).GetMember(feature.ToString())
+			.Single(x => x.DeclaringType == typeof(UnoFeature))
+			.GetCustomAttribute<UnoAreaAttribute>()?.Area;
+
+		switch (area)
+		{
+			case UnoArea.Core:
+				VerifyFeature(feature, _manifest!.UnoVersion);
+				break;
+			case UnoArea.CSharpMarkup:
+				VerifyFeature(feature, UnoCSharpMarkupVersion);
+				break;
+			case UnoArea.Extensions:
+				VerifyFeature(feature, UnoExtensionsVersion);
+				break;
+			case UnoArea.Theme:
+				VerifyFeature(feature, UnoThemesVersion);
+				break;
+			case UnoArea.Toolkit:
+				VerifyFeature(feature, UnoToolkitVersion);
+				break;
+		}
+	}
+
+	private void VerifyFeature(UnoFeature feature, string? version, [CallerArgumentExpression(nameof(version))] string? versionName = null)
+	{
+		if (string.IsNullOrEmpty(version))
+		{
+			Log.LogError(subcategory: "",
+				errorCode: "UNOB0006",
+				helpKeyword: null,
+				helpLink: "https://aka.platform.uno/UNOB0006",
+				file: null,
+				lineNumber: 0,
+				columnNumber: 0,
+				endLineNumber: 0,
+				endColumnNumber: 0,
+				message: $"The UnoFeature '{feature}' was selected, but the property {versionName} was not set.");
+		}
+	}
+
+	private void AddPackage(string packageId, string? excludeAssets = null)
+	{
+		// 1) Check for Existing References
+		var existingReference = PackageReferences.SingleOrDefault(x => x.ItemSpec == packageId);
+		if (existingReference is not null)
+		{
+			// 1.1) Validate it has a version available
+			if (PackageVersions.Any(x => x.ItemSpec == existingReference.ItemSpec) || !string.IsNullOrEmpty(existingReference.GetMetadata("Version"))
+				|| !string.IsNullOrEmpty(existingReference.GetMetadata("VersionOverride")))
+			{
+				// 1.2) Add the PackageId to the ExistingReferences so that we can log a warning at the end.
+				_existingReferences.Add(packageId);
+				return;
+			}
+
+			Log.LogWarning("The Package '{0}' has an existing PackageReference with no Version attribute or associated PackageVersion. The Uno.Sdk is removing this and adding an implicit reference.", packageId);
 			return;
 		}
 
-		AddPackage("Uno.Toolkit.WinUI", UnoToolkitVersion);
-		AddPackageForFeature(UnoFeature.Cupertino, "Uno.Toolkit.WinUI.Cupertino", UnoToolkitVersion);
-		if (HasFeature(UnoFeature.Material))
+		// 2) Load the Version from the PackageManifest. This will get the version whether it was set through MSBuild or the bundled packages.json
+		var version = _manifest!.GetPackageVersion(packageId);
+
+		// 3) Validate the version has a value. If not attempt to get the latest version from NuGet.org
+		if (string.IsNullOrEmpty(version))
 		{
-			AddPackage("Uno.Toolkit.WinUI.Material", UnoToolkitVersion);
-			AddPackageForFeature(UnoFeature.CSharpMarkup, "Uno.Toolkit.WinUI.Material.Markup", UnoToolkitVersion);
+			Log.LogWarning("The package '{0}' has no available version.", packageId);
+			using var client = new NuGetApiClient();
+			var isUnoPreview = _unoVersion?.IsPreview ?? false;
+			var preview = packageId.StartsWith("Uno.", StringComparison.InvariantCulture) && isUnoPreview;
+			version = client.GetVersion(packageId, preview);
+			Log.LogMessage(MessageImportance.High, "Retrieved the latest package version '{0}' for the package '{1}'.", version, packageId);
 		}
 
-		AddPackageForFeature(UnoFeature.CSharpMarkup, "Uno.Toolkit.WinUI.Markup", UnoToolkitVersion);
-		AddPackageForFeature(UnoFeature.Skia, "Uno.Toolkit.Skia.WinUI", UnoToolkitVersion);
-	}
-
-	public void AddUnoThemes()
-	{
-		if (HasFeature(UnoFeature.Material))
+		if (version is null || string.IsNullOrEmpty(version))
 		{
-			AddPackage("Uno.Material.WinUI", UnoThemesVersion);
-			AddPackageForFeature(UnoFeature.CSharpMarkup, "Uno.Material.WinUI.Markup", UnoThemesVersion);
-			AddPackageForFeature(UnoFeature.CSharpMarkup, "Uno.Themes.WinUI.Markup", UnoThemesVersion);
-		}
-		else
-		{
-			AddPackageForFeature(UnoFeature.Cupertino, "Uno.Cupertino.WinUI", UnoThemesVersion);
-		}
-	}
-
-	public void AddPrism()
-	{
-		if (!HasFeature(UnoFeature.Prism))
-		{
+			Debug("Unable to locate package version for '{0}'.", packageId);
 			return;
 		}
 
-		AddPackageWhen(!IsExecutable, "Prism.Uno.WinUI", PrismVersion);
-		AddPackageWhen(IsExecutable, "Prism.DryIoc.Uno.WinUI", PrismVersion);
-		AddPackageForFeature(UnoFeature.CSharpMarkup, "Prism.Uno.WinUI.Markup", PrismVersion);
+		// 4) Ensure there is not already an existing Implicit Reference that was added (this shouldn't happen)
+		var existing = _implicitPackages.SingleOrDefault(x => x.PackageId == packageId);
+		if (existing is not null)
+		{
+			Debug("An existing Implicit Package reference has already been added for '{0}'.", packageId);
+			return;
+		}
+
+		// 5) Add the Implicit Package Reference
+		Debug("Adding Implicit Reference for '{0}' with version: '{1}'.", packageId, version);
+		_implicitPackages.Add(new PackageReference(packageId, version, excludeAssets));
+	}
+
+	private void Debug(string message, params object[] args)
+	{
+		var importantance = SdkDebugging ? MessageImportance.High : MessageImportance.Low;
+
+		Log.LogMessage(importantance, message, args);
 	}
 }
